creating Router.jsx file to handle routing, import it in react app.jsx
